Workarounded bug causing some features not reported properly when landed in b47 aircraft.
It is a difficult bug related to the objects drawn when drawing the map. The list of objects is used to determine on which object this aircraft is landed. Somehow, because this aircraft models lowest z coordinate was 0, the objects on which it landed (helipads) were not reported properly. It has been solved by translating the model -0.5 on the z axis, and correcting it by adding a 0.5 belly_height. I was unable to trace back the real source of this, but it seems the only aircraft suffering it.