Joystick controllers: Update to libSDL2.
libSDL1.2 is quite deprecated and cannot even detect two joysticks at once if
they provide the events API. Switching to libSDL2 seems to fix this problem.

Additionally, this cleans up the joystick connection-type handling and
priority, as it was all unused since we switched to SDL. The menus have been
updated accordingly.