Only notify Landing to Mission when stopped
The process of figuring out if the aircraft has touched down on a
mission-related object is very expensive. Currently this was done every time
we touch-down, and it was done again when speed is 0.

This commit introduces a new specific Parked Callback that gets called when
speed is 0 and moves the relevant "check that we are at the right place" to
it.

Thus, the Touch-Down-Callback becomes cheaper. This removes most the
lagging on touch down, which is specially visible with the F22 (which
normally lands with speed > 0, unlike helicopters in most cases).

This adds to a series of improvements
(48fc3d70443a0906361abef7b3258b7fc9c47947 and
37f45d56609e70b4a9a1da9a19a257729293b638) that caused really bad
lagging when landed (or landing).