Fix #44: ghosting on intel. GL_LINES_ADJACENCY errors.
See added comments in sardrawrunway.c.

Basically there is an opengl error message on my intel hardware worst of all
some runway elements are ghosting all over the screen as soon as we run a
ground contact check (which switches gl rendering mode to GL_SELECT for a
frame).

This workaround disables rendering those elements during GL_SELECT. It seems
to work mostly fine. I'm unsure that the issue is in the game but who knows.